Revert "Pull out `MockAccessibilityManager` from `test_mousekey_config.cpp`"
This reverts commit 6e7211dd0993a46fb4bbd5032a90ce121d84391f.
